+++ plugins/core/src/components/plugin/PluginPreview.tsx
@@ -1,6 +1,7 @@
 import { Alert, AlertVariant } from '@patternfly/react-core';
 import React, { useContext } from 'react';
 
+import { AuthContext, IAuthContext } from '../../context/AuthContext';
 import { IPluginPreviewProps, IPluginsContext, PluginsContext } from '../../context/PluginsContext';
 
 export const PluginPreview: React.FunctionComponent<IPluginPreviewProps> = ({
@@ -9,6 +10,8 @@ export const PluginPreview: React.FunctionComponent<IPluginPreviewProps> = ({
   name,
   options,
 }: IPluginPreviewProps) => {
+  const authContext = useContext<IAuthContext>(AuthContext);
+
   const pluginsContext = useContext<IPluginsContext>(PluginsContext);
   const pluginDetails = pluginsContext.getPluginDetails(name);
   const Component =
@@ -16,7 +19,7 @@ export const PluginPreview: React.FunctionComponent<IPluginPreviewProps> = ({
       ? pluginsContext.components[pluginDetails.type].preview
       : undefined;
 
-  if (!pluginDetails || !Component) {
+  if (!pluginDetails || !Component || !authContext.hasPluginAccess(pluginDetails.name)) {
     return (
       <Alert variant={AlertVariant.danger} isInline={true} title="Plugin was not found">
         {pluginDetails ? (